% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigR.R
\name{add.image}
\alias{add.image}
\title{add.image}
\usage{
add.image(self, image, name)
}
\arguments{
\item{self}{- An object of class FigR}

\item{image}{- The ggplot object to add.}

\item{name}{- The name to give the plot object}
}
\value{
Returns null
}
\description{
- Adds a plot object to the FigR container.
}

