% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tools.R
\name{cap}
\alias{cap}
\title{cap}
\usage{
cap(x, lbnd = -Inf, ubnd = Inf)
}
\arguments{
\item{x}{- Input}

\item{lbnd}{- The lower bound. Defaults to -Inf}

\item{ubnd}{- The upper bound. Defaults to Inf}
}
\value{
Returns the input value if it's within the bounds. If the input
exceeds one of the bounds, cap returns the bound value instead.
}
\description{
Imposes lower and upper bounds on a value.
}
