% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tools.R
\name{cat.cap}
\alias{cat.cap}
\title{cat.cap}
\usage{
cat.cap(vector, lim = 50, ret = "other")
}
\arguments{
\item{vector}{- The input vector}

\item{lim}{- A numeric limit on the number}

\item{ret}{- The category that all other values are collapsed into. Defaults
to 'other'.}
}
\value{
Returns the vector, where all values not in the x top values are
replaced with the value for ret.
}
\description{
Takes a vector of categorical values and collapses the vector
into the first 'lim' unique values (by volume), and the rest are collapsed to
an 'other' category.
}

