% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looker.tools.R
\name{clean_looker_colnames}
\alias{clean_looker_colnames}
\title{clean_looker_colnames}
\usage{
clean_looker_colnames(df)
}
\arguments{
\item{df}{The dataframe being altered.}
}
\value{
Returns the data frame with the new snake case column names.
}
\description{
Transforms looker's standard formatting into a much more
data-friendly snake case format. NOTE: The column names must already be
formatted according to the URL formatting used to call the query. Looker3's
pretty formatting will fail. To work around this, replace the column names
with the vector you specified in the dictionary used to call the query.
}

