% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tools.R
\name{clip.tail}
\alias{clip.tail}
\title{clip.tail}
\usage{
clip.tail(x, repl = 0)
}
\arguments{
\item{x}{- a vector whose tail you'll be clipping.}

\item{repl}{- The value to replace. Defaults to zero. #'}
}
\value{
- Returns a vector where any continuous series of repl at the end of
the vector is replace with NA
}
\description{
clip.tail
}

