% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tools.R
\name{\%||\%}
\alias{\%||\%}
\title{%||%}
\usage{
x \%||\% y
}
\value{
equivalent to postgres coalesce(x,y)
}
\description{
DataSci's coalesce function. Gives postgres-like coalesce
}
\examples{
NULL \%||\% 'a'
}

