% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tools.R
\name{corr.plot}
\alias{corr.plot}
\title{corr.plot}
\usage{
corr.plot(df)
}
\arguments{
\item{df}{data.frame. A dataframe whose features you want to inspect the 
correlations of.}
}
\value{
A graph object from igraph::dataframe. Can be fed straight into plot.
The plot pushes more heavily correlated variables closer to each other and
less correlated variables further apart. Useful for evaluating potential
factor structures in large data sets.
}
\description{
corr.plot
}
