% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tools.R
\name{cum.sum}
\alias{cum.sum}
\title{cum.sum}
\usage{
cum.sum(x)
}
\arguments{
\item{x}{- A numerical vector}
}
\value{
- Returns a cumulative sum vector that is more suitable to graphing.
}
\description{
- A better cumulative summing function. This one ignores
NA/Inf values in the cumulative sum's history, but also returns NA for
the values that are explicitly NA/Inf. The result noticeable graphically -
Troublesome values are treated as missing, but the missing value itself
does not cause a problem when computing the prior history.
}

