% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tools.R
\name{data.frame.compare}
\alias{data.frame.compare}
\title{data.frame.compare}
\usage{
data.frame.compare(df, df2, ...)
}
\arguments{
\item{df}{- The first data frame}

\item{df2}{- The second data frame}

\item{by}{- Used in merge(). The list of column names to join by. The default
on merge is all same-named columns, which would defeat a lot of the
functionality here. So the user must define the join explicitly.}
}
\value{
A list object containing
}
\description{
Compares two data frames by joining. Returns a summary containing
duplicate, missing, and differing records.
}

