% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigR.R
\name{delete.image}
\alias{delete.image}
\title{delete.image}
\usage{
delete.image(self, name.or.id)
}
\arguments{
\item{self}{- An object of class FigR}

\item{name.or.id}{- Either the name of the object, or its id.}
}
\value{
Returns null
}
\description{
- Removes a plot object to the FigR container.
}

