% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.plot.R
\name{.buckets}
\alias{.buckets}
\title{.buckets}
\usage{
.buckets(x, num.switch = class(x) \%in\% c("numeric", "integer",
  "float"))
}
\arguments{
\item{x}{- A vector to be bucketed.}

\item{num.switch}{- Binary. True if we should treat x as numeric, false if
categorical.}
}
\value{
Returns a list object with 2 values. $groups is a vector containing
the distinct bucket values. $buckets returns the snapped values.
}
\description{
- Does the automatic bucketing for the plot. Really just
cribbing on hist$breaks to do the heavy work, but with some better handling
of categorical variables. Also includes a switch to treat anything numeric
as categorical instead.
}
\examples{
.buckets(iris[,1], F)
.buckets(iris[,2])
}
