% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tools.R
\name{empty.frame}
\alias{empty.frame}
\title{empty.frame}
\usage{
empty.frame(col.names)
}
\arguments{
\item{col.names}{- User input column names.}
}
\value{
An empty data frame object with the specified column names
}
\description{
Create an empty data frame. Surprisingly tricky to do.
}
