% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.tools.R
\name{env.size}
\alias{env.size}
\title{Title env.size}
\usage{
env.size(env)
}
\arguments{
\item{env}{- The name of the environment, as a string.}
}
\description{
Get the size of an environment by typing in its name as as string.
}
