% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigR.R
\name{FigR}
\alias{FigR}
\title{FigR}
\usage{
FigR(write.path = getwd(), index = 1, type = ".png")
}
\arguments{
\item{write.path}{- Path for writing out all the collected images. Defaults
to the working directory.}

\item{index}{- FigR will append the phrase 'figXXX' to the name of the file.
Index determines the value the starting value. So the first three files
printed out by FigR will have names 'fig001...', 'fig002...', and 'fig003...'.
Defaults to 1}

\item{type}{- file type to save as, as text. Defaults to .png.}
}
\description{
Class constructor for image handling. FigR has a path value
telling it where to print out, a method to add an image, a method to delete
an image, and a method to print out all images to the path.
}

