% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.tools.R
\name{file.paths}
\alias{file.paths}
\title{These two will enumerate the .csv files in a path, as well as load them
into a data frame.}
\usage{
file.paths(x)
}
\description{
These two will enumerate the .csv files in a path, as well as load them
into a data frame.
}

