% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.tools.R
\name{find.dates}
\alias{find.dates}
\title{Will parse query text and returns base on return.place. If true
function returns a vector of the available dates in a query. If 
false, returns a vector of the logical position of dates in the
vector created by query.split.}
\usage{
find.dates(q, return.place = F)
}
\description{
Will parse query text and returns base on return.place. If true
function returns a vector of the available dates in a query. If 
false, returns a vector of the logical position of dates in the
vector created by query.split.
}

