% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.conversions.R
\name{flg.create}
\alias{flg.create}
\title{This function is a wrapper for as.flgs, to apply it to a named list
of columns in a target data frame. This allows you to do batch the
flag conversion of factor variables. As above, be careful with this
function, as it's easy to flood memory by batching more than one
column with large numbers of unique entries.}
\usage{
flg.create(df, name.list)
}
\description{
This function is a wrapper for as.flgs, to apply it to a named list
of columns in a target data frame. This allows you to do batch the
flag conversion of factor variables. As above, be careful with this
function, as it's easy to flood memory by batching more than one
column with large numbers of unique entries.
}

