% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.tools.R
\name{get.files}
\alias{get.files}
\title{get.files}
\usage{
get.files(x)
}
\arguments{
\item{x}{- a single path, expressed as a character string}
}
\value{
Function returns a single dataframe, which is the appended result
of all .csv files at the path x.
}
\description{
Takes an input file path, look for all .csv files, and then
append each file in that path into a single data frame.
}

