% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tools.R
\name{impute.val}
\alias{impute.val}
\title{impute.var}
\usage{
impute.val(df, var, formula.x, plots = F)
}
\arguments{
\item{df}{The data frame.}

\item{var}{The name of the variable to impute null values for, as a string}

\item{formula.x}{The formula for the imputation model}

\item{plots}{A variable to turn diagnostic plots on or off.}
}
\value{
Returns the original data frame with a new variable labeled "imp_" +
var. Designed to put into magrittry chains easily.
}
\description{
Creates a quick linear model to impute null values based on
values in the data. Will only replace nulls, the predicted model values are
discarded.
}

