% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.tools.R
\name{is.date.text}
\alias{is.date.text}
\title{is.date.text}
\usage{
is.date.text(x)
}
\arguments{
\item{x}{- An input string}
}
\value{
Function returns logical, to indicate whether the string could be
turned into a date.
}
\description{
is.date.text
}

