% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.tools.R
\name{knitrdone}
\alias{knitrdone}
\title{Knitrdone}
\usage{
knitrdone(dir, ...)
}
\arguments{
\item{dir}{The directory to write the output doc and all intermediaries to.}

\item{...}{Additional parameters can be passed to ezknit.}
}
\value{
Returns the .html and .md documents.
}
\description{
Creates a wrapper around ezknit, which allows knitting to a
specific output directory without messing up the working directory /
environment varialbes. This wrapper just creates directories if none exist,
and it can be placed at the beginning of a .RMD doc as the custom knit
function.
}
