% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.explorer.R
\name{look.for}
\alias{look.for}
\title{look.for}
\usage{
look.for(list.list, target)
}
\arguments{
\item{target}{- name of the attribute.}

\item{list}{- a list of lists}
}
\value{
a single vector containing the value of the t
}
\description{
Looks for a single named attribute in a list of lists. Returns a
vector containing every instance of that attribute. Pretty handy for playing
with API call results.
}

