% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looker.tools.R
\name{looker.chunker}
\alias{looker.chunker}
\title{looker.chunker}
\usage{
looker.chunker(looker.path, data.path, start.date, end.date,
  interval = "1 month")
}
\arguments{
\item{looker.path}{- The path to the file that contains the looker dictionary}

\item{data.path}{- The path to your data dump}

\item{start.date}{- The start date of the query}

\item{end.date}{- The end date of the query.}

\item{interval}{- String input for the interval size. See the 'by' input
parameter from seq()}
}
\value{
Returns null. The function writes out the data frames to data.path.
}
\description{
This function takes a looker query and chops it up into smaller
portions, ideally to improve runtime. The looker dictionary must only have
two date fields, and they must be sequentially ordered so start.date comes
first.
}
