% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.explorer.R
\name{name.search}
\alias{name.search}
\title{name.search}
\usage{
name.search(x, top = "")
}
\arguments{
\item{x}{- Any object}
}
\value{
names(x), but iterated recursively through every level of the object
names.
}
\description{
Iterates through the list of attributes of an object. Unlike
calling attributes, name.search works recursively. Really great for xml/html
returns where you want to get a sense of the full hierarchy.
}

