% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.plot.R
\name{perf.table}
\alias{perf.table}
\title{perf.plot}
\usage{
perf.table(y, x, num.switch = class(x) \%in\% c("numeric", "integer",
  "float"))
}
\arguments{
\item{y}{- The target variable}

\item{x}{- The independent variable.}
}
\value{
Returns a dataframe with the mean value of y at bucketed values of x.
}
\description{
Creates a table for a common univariate analysis - the target variable
averaged at different levels of x. Table returns mean values and counts at the
different levels of x, where the levels are determined by using hist()$breaks.
}
