% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homemade.themes.R
\name{qsave}
\alias{qsave}
\title{qsave}
\usage{
qsave(filename, height = 6, width = 9, dpi = 600, ...)
}
\arguments{
\item{filename}{- file name/filename of plot}

\item{height}{- ggsave input. Defaults to 6.}

\item{width}{- ggsave input. Defaults to 9.}

\item{dpi}{- ggsave input. Defaults to 600.}
}
\description{
Just a ggsave wrapper but with presaved settings
}

