% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.tools.R
\name{query.chunker}
\alias{query.chunker}
\title{query.chunker}
\usage{
\method{query}{chunker}(query.path, data.path, start.date, end.date,
  interval = "1 month")
}
\arguments{
\item{query.path}{- The path to the query, as a string.}

\item{data.path}{- The path to dump the data into, as a string}

\item{start.date}{- The start date for the query.}

\item{end.date}{- The end date for the query.}

\item{interval}{- A string input for the interval to chunk by. See 'by' input
parameter from seq(). Defaults to '1 month'.}
}
\description{
This function chunks a query into smaller portions, to improve
runtime performance and get around that 2 hour forced limit on the DB. The
query must only contain two text strings formatted as a date; if there are
additional hardcoded dates, this function will break.
}

