% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.tools.R
\name{read.query}
\alias{read.query}
\title{#A wrapper for read_query, with some additional workarounds for 
the way Windows handles text files. 
TODO: This function collapses queries in a way that's harmful to
comments made via --.}
\usage{
read.query(x)
}
\description{
#A wrapper for read_query, with some additional workarounds for 
the way Windows handles text files. 
TODO: This function collapses queries in a way that's harmful to
comments made via --.
}

