% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.tools.R
\name{replace.dates}
\alias{replace.dates}
\title{Replaces query dates with a vector of dates specified by the user.
Function will error if length(dates) is not the same as dates found
by find.dates}
\usage{
replace.dates(q, dates)
}
\description{
Replaces query dates with a vector of dates specified by the user.
Function will error if length(dates) is not the same as dates found
by find.dates
}

