% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.tools.R
\name{run.query}
\alias{run.query}
\title{run.query}
\usage{
run.query(query, conn = last_connection())
}
\description{
is exactly berdie::read_query. I just like having function
names separated by .s and column names separated by _
}

