% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tools.R
\name{s.cap}
\alias{s.cap}
\title{s.cap}
\usage{
s.cap(v, lbnd = -Inf, ubnd = Inf)
}
\arguments{
\item{v}{- A vector input.}

\item{lbnd}{- The lower bound. Defaults to -Inf.}

\item{ubnd}{- The upper bound. Defaults to Inf}
}
\value{
Returns the vector, where values are replaced by the bound if they
exceed that limit.
}
\description{
An sapply wrapper for cap.
}

