% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.tools.R
\name{s3.read}
\alias{s3.read}
\title{Title s3read}
\usage{
s3.read(bucket, filepath)
}
\arguments{
\item{bucket}{The top-level bucket the data is stored in.}

\item{filepath}{The remaining path to the file}
}
\value{
Returns the serialized R object.
}
\description{
A windows-based s3read command. Designed to mimic the function
of robertzk/s3mpi, but using windows system commands.
}
