% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.tools.R
\name{s3.write}
\alias{s3.write}
\title{Title s3write}
\usage{
s3.write(obj, bucket, filepath)
}
\arguments{
\item{bucket}{The top-level bucket the data is stored in.}

\item{filepath}{The full path where you'll store the file}
}
\description{
A windows-based s3write command. Designed to mimic the function
of robertzk/s3mpi, but using windows system commands.
}
