% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.tools.R
\name{s3bucketlist}
\alias{s3bucketlist}
\title{Title s3bucketlist}
\usage{
s3bucketlist()
}
\value{
Returns a data frame listing the available s3 buckets and some
attributes
}
\description{
Calls all of the available s3 buckets.
}

