% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.tools.R
\name{s3list}
\alias{s3list}
\title{s3list}
\usage{
s3list(bucket, prefix = "", lim = if (prefix == "") {     20 } else {    
  10000 })
}
\arguments{
\item{bucket}{The bucket we'll be searching}

\item{prefix}{Optional to restrict your search to a specific directory path.}

\item{lim}{Optional}
}
\description{
s3list
}

