% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.tools.R
\name{snake.case}
\alias{snake.case}
\title{snake.case}
\usage{
snake.case(text)
}
\arguments{
\item{text}{- User Input Text}
}
\value{
Function returns text that is formatted to snake case.
}
\description{
snake.case
}
