% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.tools.R
\name{source.all}
\alias{source.all}
\title{source.all}
\usage{
source.all(path)
}
\arguments{
\item{path}{- Path to the directory you wish to source}
}
\value{
returns null
}
\description{
Why source one file when you can source a whole directory?
}

