% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homemade.themes.R
\name{theme_BA}
\alias{theme_BA}
\title{theme_BA()}
\usage{
theme_BA()
}
\value{
ggplot theme object, based off ggthemes::theme_gdocs()
}
\description{
Creates my go-to theme for the ggplot in analyses.
}

