% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tools.R
\name{toss.tail}
\alias{toss.tail}
\title{toss.tail}
\usage{
toss.tail(x, n)
}
\arguments{
\item{x}{- a vector whose tail you'll be tossing.}

\item{n}{- the number of observations to replace with null.}
}
\value{
- Returns a vector where the last n non-NA values are replaced with
NA.
}
\description{
toss.tail
}
