% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.sum.R
\name{var.sum}
\alias{var.sum}
\title{This function takes a data frame as an input and spits out a summary data 
frame.That frame lists out the following features: total count, class, 
null count, unique count, and a character string of the first couple values}
\usage{
var.sum(df)
}
\description{
This function takes a data frame as an input and spits out a summary data 
frame.That frame lists out the following features: total count, class, 
null count, unique count, and a character string of the first couple values
}

