% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.tools.R
\name{vec.chunker}
\alias{vec.chunker}
\title{vec.chunker}
\usage{
vec.chunker(x, n = 100)
}
\arguments{
\item{x}{- any vector}
}
\value{
Returns a list
}
\description{
Splits a vector into a list where each list element contains
the next n values from the vector.
}

