% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.tools.R
\name{vec.text}
\alias{vec.text}
\title{vec.text}
\usage{
vec.text(x)
}
\arguments{
\item{x}{- any character vector}
}
\value{
A text string formatted like an R vector, ie "c(var1, var2, ...)"
}
\description{
Turns a vector into the string that would represent that vector.
For when you need to skip the bullshit and hardcode a list into the code.
}

