% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigR.R
\name{write.all}
\alias{write.all}
\title{write.all}
\usage{
write.all(self)
}
\arguments{
\item{self}{- An object of class FigR}
}
\value{
Returns null
}
\description{
- Takes the FigR object and writes out all elements in
$images to the directory specified by $write.path.
}

