% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tools.R
\name{zero.out}
\alias{zero.out}
\title{zero.out}
\usage{
zero.out(x)
}
\arguments{
\item{x}{- A numeric vector.}
}
\value{
Returns x but with NA/Inf values replaced with zero.
}
\description{
Replaces NA/Inf values with 0.
}
