% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.conversions.R
\name{as.flgs}
\alias{as.flgs}
\title{This function converts a single column into a matrix of binary
flags, one for each unique value the column can take. The columns
can then be readily joined with the original data frame using cbind.
Be careful with this function at present, as you can easily flood 
memory by choosing columns with large numbers of unique values.}
\usage{
as.flgs(df, x)
}
\description{
This function converts a single column into a matrix of binary
flags, one for each unique value the column can take. The columns
can then be readily joined with the original data frame using cbind.
Be careful with this function at present, as you can easily flood 
memory by choosing columns with large numbers of unique values.
}

