% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.tools.R
\name{file.paths}
\alias{file.paths}
\title{file.paths}
\usage{
file.paths(x)
}
\arguments{
\item{x}{- a single path, expressed as a character string}
}
\value{
Function returns a vector of the global paths for each file found
in x.
}
\description{
Takes an input file path and returns the full system path
for each file found at the input path.
}

