% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tools.R
\name{get.files}
\alias{get.files}
\title{get.files}
\usage{
get.files(x, match = "\\\\.csv$")
}
\arguments{
\item{x}{- a single path, expressed as a character string}

\item{match}{- A user-input regex to match to a subset of the filenames in a
directory. Will only get files where grep(match)=T. The default value matches
any .csv}
}
\value{
Function returns a single dataframe, which is the appended result
of all .csv files at the path x.
}
\description{
Takes an input file path, look for all .csv files, and then
append each file in that path into a single data frame.
}
