% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.tools.R
\name{replace.dates}
\alias{replace.dates}
\title{replace.dates}
\usage{
replace.dates(text, dates, date.reg = "\\\\d+[/-]\\\\d+[/-]\\\\d+",
  text.split = function(x) x \%>\% as.character \%>\% strsplit(" ") \%>\%
  unlist, text.combine = function(x) paste(x, collapse = " "))
}
\arguments{
\item{text}{- query as a single character string}

\item{dates}{- a vector of dates to replace in the query. Function will error
if length(dates) != length(find.dates(text,T))}
}
\value{
Function returns the original query text, but with the new dates.
}
\description{
Replaces query dates with a vector of dates specified by the
user. Function will error if length(dates) is not the same as dates found by
find.dates.
}
