% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.tools.R
\name{run.query}
\alias{run.query}
\title{run.query}
\usage{
run.query(x, conn = reset.conn())
}
\arguments{
\item{x}{- query as a character vector}

\item{conn}{- as the connection object. See berdie::postgresql_connection}
}
\description{
is exactly berdie::read_query. I just like having function
names separated by .s and column names separated by _

Edit: That's how this started. Then I realized that given how the average
query length is 20 mins longish and the DB has a tendency to reset itself
after a couple, there's no reason not to make conn default to a new
connection every time. Hopefully the DB is good at dropping connections.
Trololololol.
}

