% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tools.R
\name{var.sum}
\alias{var.sum}
\title{var.sum}
\usage{
var.sum(df, exclude = "values")
}
\arguments{
\item{df}{- A dataframe}

\item{exclude}{-}
}
\value{
A dataframe that summarizes the data, including total count, # nulls
# uniques, variable class, and text showing some examples.
}
\description{
This function takes a data frame as an input and spits out a summary data
frame.That frame lists out the following features: total count, class,
null count, unique count, and a character string of the first couple values
}
\examples{
var.sum(iris)
}
